[rtl872x] hal: implement i2c sleep and fix the reset APIs

+++ hal/inc/i2c_hal.h
@@ -105,7 +105,7 @@ void hal_i2c_flush(hal_i2c_interface_t i2c, void* reserved);
 bool hal_i2c_is_enabled(hal_i2c_interface_t i2c, void* reserved);
 void hal_i2c_set_callback_on_received(hal_i2c_interface_t i2c, void (*function)(int), void* reserved);
 void hal_i2c_set_callback_on_requested(hal_i2c_interface_t i2c, void (*function)(void), void* reserved);
-uint8_t hal_i2c_reset(hal_i2c_interface_t i2c, uint32_t reserved, void* reserve1);
+int hal_i2c_reset(hal_i2c_interface_t i2c, uint32_t reserved, void* reserve1);
 int hal_i2c_sleep(hal_i2c_interface_t i2c, bool sleep, void* reserved);
 int32_t hal_i2c_lock(hal_i2c_interface_t i2c, void* reserved);
 int32_t hal_i2c_unlock(hal_i2c_interface_t i2c, void* reserved);
